# src/app/docs/swagger-audiocore.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Audio Core API",
    "description": "Advanced audio generation system. Powered by Suno AI and ChatGPT.",
    "version": "1.2.0"
  },
...
  "paths": {
...
    "/api/custom_generate": {
      "post": {
        "summary": "Custom Generation",
        "description": "Generate music with custom lyrics, style tags, and title. Optionally use a persona (voice profile) for vocal consistency.",
        "tags": [
          "Generation"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "prompt",
                  "tags",
                  "title"
                ],
                "properties": {
                  "prompt": {
                    "type": "string",
                    "description": "Lyrics for the song.",
                    "example": "[Verse 1]\nNeon lights in the rain..."
                  },
                  "tags": {
                    "type": "string",
                    "description": "Musical style tags.",
                    "example": "synthwave neon nocturnal"
                  },
                  "title": {
                    "type": "string",
                    "description": "Title for the song.",
                    "example": "Midnight City"
                  },
                  "make_instrumental": {
                    "type": "boolean",
                    "description": "Generate instrumental only (default: false)",
                    "default": false,
                    "example": false
                  },
                  "model": {
                    "type": "string",
                    "description": "Model version (default: chirp-crow)",
                    "default": "chirp-crow",
                    "example": "chirp-crow"
                  },
                  "wait_audio": {
                    "type": "boolean",
                    "description": "Wait for generation to complete (default: false)",
                    "default": false,
                    "example": true
                  },
                  "negative_tags": {
                    "type": "string",
                    "description": "Style elements to avoid",
                    "example": "slow, acoustic"
                  },
                  "persona_id": {
                    "type": "string",
                    "description": "Optional persona ID for vocal consistency. Requires artist_clip_id.",
                    "example": "ce93a372-48f8-4ff7-aaa9-d57f871b75dd"
                  },
                  "artist_clip_id": {
                    "type": "string",
                    "description": "Original clip ID the persona was created from. Required when using persona_id. You can also use 'root_clip_id' instead.",
                    "example": "aca95097-79d7-4120-b9b7-42a9cdc07ce2"
                  },
                  "root_clip_id": {
                    "type": "string",
                    "description": "Alias for artist_clip_id. Use the root_clip_id value from /api/personas response.",
                    "example": "aca95097-79d7-4120-b9b7-42a9cdc07ce2"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/audio_info"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "402": {
            "description": "Insufficient credits"
          }
        }
      }
    },
...
    "/api/get_wav": {
      "get": {
        "summary": "Download WAV Audio (Pro)",
        "description": "Get the WAV file URL for a song. This is a Pro feature that provides lossless audio quality. The endpoint initiates conversion and waits for the WAV to be ready.",
        "tags": [
          "Management"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": { "type": "string" },
            "description": "The song ID to download as WAV",
            "example": "1945cb16-81d6-4448-b8d2-4a73841abb0a"
          }
        ],
        "responses": {
          "200": {
            "description": "WAV URL retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "wav_url": {
                      "type": "string",
                      "description": "Direct URL to download the WAV file",
                      "example": "https://cdn1.suno.ai/1945cb16-81d6-4448-b8d2-4a73841abb0a.wav"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required parameter: id"
          },
          "403": {
            "description": "Pro subscription required"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/generate_stems": {
      "post": {
        "summary": "Generate Stems from Audio",
        "description": "Extract isolated audio stems (vocals, instrumental, etc.) from a complete song. Returns stem IDs that can be used to download the separated tracks.",
        "tags": [
          "Generation"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["audio_id"],
                "properties": {
                  "audio_id": {
                    "type": "string",
                    "description": "ID of the song to extract stems from",
                    "example": "1945cb16-81d6-4448-b8d2-4a73841abb0a"
                  },
                  "stem_type": {
                    "type": "string",
                    "enum": ["two", "all"],
                    "default": "two",
                    "description": "Type of stem extraction: 'two' for Vocals+Instrumental (10 credits), 'all' for all detected stems (50 credits)",
                    "example": "two"
                  },
                  "wait": {
                    "type": "boolean",
                    "default": false,
                    "description": "If true, waits for stem generation to complete before returning",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stems generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "description": "Stem clip ID"
                      },
                      "title": {
                        "type": "string",
                        "description": "Stem title (e.g., 'Lead Vocals', 'Instrumental')"
                      },
                      "status": {
                        "type": "string",
                        "description": "Generation status"
                      },
                      "audio_url": {
                        "type": "string",
                        "description": "URL to MP3 file (when complete)"
                      },
                      "stem_from_id": {
                        "type": "string",
                        "description": "Original song ID"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "402": {
            "description": "Insufficient credits"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/download_stems": {
      "post": {
        "summary": "Download Stems as WAV (Pro)",
        "description": "Generate stems from a song and download them as WAV files. This is a Pro feature that combines stem generation with WAV conversion. Returns WAV URLs for each stem.",
        "tags": [
          "Management"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["audio_id"],
                "properties": {
                  "audio_id": {
                    "type": "string",
                    "description": "ID of the song to extract stems from",
                    "example": "1945cb16-81d6-4448-b8d2-4a73841abb0a"
                  },
                  "stem_type": {
                    "type": "string",
                    "enum": ["two", "all"],
                    "default": "two",
                    "description": "Type of stem extraction: 'two' for Vocals+Instrumental (10 credits), 'all' for all detected stems (50 credits)",
                    "example": "two"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stems downloaded as WAV successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "description": "Stem clip ID"
                      },
                      "title": {
                        "type": "string",
                        "description": "Stem title"
                      },
                      "wav_url": {
                        "type": "string",
                        "description": "URL to download WAV file",
                        "example": "https://cdn1.suno.ai/stem-id.wav"
                      },
                      "audio_url": {
                        "type": "string",
                        "description": "URL to MP3 file"
                      },
                      "stem_from_id": {
                        "type": "string",
                        "description": "Original song ID"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "402": {
            "description": "Insufficient credits"
          },
          "403": {
            "description": "Pro subscription required for WAV download"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "audio_info": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "audio_url": { "type": "string" },
          "status": { "type": "string" }
        }
      },
...
    }
  }
}